+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Place a file called `.env` in either the MuMu_RichPresence Directory or it's parent
 - ⚡️Added opening the log file as a tray option if ExtendedLogging is enabled
 - 🦺 Bugfix: Toggling the Rich Presence from the tray would not re-enable the presence under certain circumstances
+- ⚡ The Setup version of the program will now log to `%LocalAppData%\PlayGames-RichPresence\`
+  - Previously logged to `%LocalAppData%\PlayGames-RichPresence\current\`
 
 ## [2.0.0] / 2025-12-23
 - ⚡️Rich Presences now have clickable links directing them to the respective game's listing on the Play Store